+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added `Vector` class
 - Added cast for Laravel
 
 ## 0.1.0 (2022-08-05)